Added shared-id to Players.
Modified tests to verify shared id is preserved on clone().
The shared ID is a ridiculous idea that I had to prevent representation
exposure, while providing a guaranteed way to determine if two Players
are intended to represent the same data. This would allow the name of a
player to be updated, probably providing a callback interface, but most
importantly it provides an easy to preserve a mapping between a
particular player and location (while allowing bad callers to fail to
update names and types on change).